refactor(foundry): move platform into new directory out of tradelens/
