+++ README.foundry.md
@@ -1,0 +1,9 @@
+<div style="text-align:center">
+    <img src="logo/tl-logo.png" width="140" height="140" alt=""/>
+</div>
+
+#### Enterprise grade solution
+
+More at: `/foundry/docs/`
+
+![foundry](https://github.com/user-attachments/assets/c2434000-27a9-4d55-9f0e-21801c45dde9)
