+++ foundry/Foundry.slnx
@@ -1,0 +1,1 @@
+<Solution />
