+++ foundry/src/Hosts/Foundry.Api/Foundry.Api.csproj
