+++ foundry/src/Hosts/Foundry.Api/Foundry.Api.http
@@ -1,0 +1,6 @@
+@Foundry.Api_HostAddress = http://localhost:5201
+
+GET {{Foundry.Api_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
