+++ foundry/src/Modules/Discussion/Foundry.Discussion.Api/Foundry.Discussion.Api.csproj
