+++ foundry/src/Modules/Discussion/Foundry.Discussion.Domain/Foundry.Discussion.Domain.csproj
