+++ foundry/src/Modules/Discussion/Foundry.Discussion.Infrastructure/Foundry.Discussion.Infrastructure.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Foundry.Discussion.Application\Foundry.Discussion.Application.csproj" />
+  </ItemGroup>
+
+</Project>
