+++ foundry/src/Shared/Foundry.SharedInfra/Foundry.SharedInfra.csproj
