+++ foundry/src/Shared/Foundry.SharedKernel/Foundry.SharedKernel.csproj
