+++ foundry/tests/Architecture/placeholder
@@ -1,0 +1,3 @@
+Enforce module and project boundaries
+
+Foundry.ArchitectureTests - NetArchTest
