--- tradelens/_src/Hosts/Tradelens.Api/Tradelens.Api.http
@@ -1,6 +1,0 @@
-@Tradelens.Api_HostAddress = http://localhost:5032
-
-GET {{Tradelens.Api_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
