--- tradelens/_src/Modules/Discussion/Tradelens.Discussion.Application/Tradelens.Discussion.Application.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Tradelens.Discussion.Domain\Tradelens.Discussion.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
