--- tradelens/_src/Shared/placeholder
@@ -1,4 +1,0 @@
-Cross-cutting Concerns - shared kernels
-
-Tradelens.SharedKernel - BaseEntity, Result<T>, ISpecification, Pagination primitives
-Tradelens.SharedInfra - ResponseCacheService, Global EF Core configurations
