--- tradelens/_tests/Architecture/placeholder
@@ -1,3 +1,0 @@
-Enforce module and project boundaries
-
-Tradelens.ArchitectureTests - NetArchTest
