--- tradelens/_tests/Hosts/placeholder
@@ -1,1 +1,0 @@
-Tradelens.Api.IntegrationTests - Test the whole API host end-to-end
