--- tradelens/_tests/Modules/placeholder
@@ -1,2 +1,0 @@
-Tradelens.Discussion.UnitTests/
-...
